* commit 'v7.7.21': (599 commits)
  Preparing for 7.7.21 release.
  scopeGuard is not backported yet.
  Fix revisionDate field definitions for several assets.  Fixes bug #11066
  drop another SQL Form table if the SQL form isn't used.  Fixes bug #11061
  Get rid of old subscription tables.  Fixes bug #11062
  fix bug #11064, spelling error in subscription i18n
  Upgrade scripts use the new API to force package installs.
  Packages used in upgrades get installed regardless of revisionDate.  Fixes bug #11059
  fixed: DataForm allows edits when locked by another version tag
  Fix a Shop bug dealing with shippers that are no longer available.  Fixes bug #11001.
  Fix EMS Schdedule template bugs.  Fixes bug #10988.
  Put package into the right place, really.
  Backport the wiki page history tab fix.
  Do not display the history tab in the wiki page if the user can't see the history page.  Fixes bug #11056
  Make sure all version tags are cleaned up.
  Add no-cache headers to maintenance screen.  Fixes bug #11037
  fixed bug where shortcuts weren't purged along with assets when both were in the trash
  fixed bug #11052; includes test
  More test cleanups.
  Clean up Sessions that are built for testing.
  ...

Conflicts:
	sbin/testEnvironment.pl

+++ lib/WebGUI/Operation/Group.pm
@@ -108,11 +108,36 @@ sub canView {
 }
 
 #-------------------------------------------------------------------
+
+=head2 doGroupSearch ($session, $op, $returnPaginator, $groupFilter)
+
+Do a search of group names and descriptions for the keyword set in the
+form variable C<groupSearchKeyword>.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=head3 $op
+
+A URL query fragment to use with the paginator.
+
+=head3 $returnPaginator
+
+If set to true, then this function will return a WebGUI::Paginator
+object.
+
+=head3 $groupFilter
+
+An array reference of groupIds to exclude from the search.
+
+=cut
+
 sub doGroupSearch {
-	my $session = shift;
-	my $op = shift;
+	my $session         = shift;
+	my $op              = shift;
         my $returnPaginator = shift;
-        my $groupFilter = shift;
+        my $groupFilter     = shift || [];
         push(@{$groupFilter},0);
         my $keyword = $session->scratch->get("groupSearchKeyword");
         if ($session->scratch->get("groupSearchModifier") eq "startsWith") {
@@ -137,6 +162,26 @@ sub doGroupSearch {
 
 
 #-------------------------------------------------------------------
+
+=head2 getGroupSearchForm ($session, $op, $params)
+
+Build and render a form for doing group searching.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=head3 $op
+
+The operation that this form should call when submitted.
+
+=head3 $params
+
+A hashref of hidden form parameters and values to add to the form.
+
+
+=cut
+
 sub getGroupSearchForm {
 	my $session = shift;
 	my $op = shift;
@@ -188,6 +233,29 @@ sub getGroupSearchForm {
 
 
 #-------------------------------------------------------------------
+
+=head2 walkGroups ($session, $parentId, $indent)
+
+Recursively find all groups which are members of this group.  Each
+group is rendered with an indent, a delete from grouping icon and
+an edit icon.
+
+Returns the HTML.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=head3 $parentId
+
+The GUID of the group to display the children of.
+
+=head3 $indent
+
+A snippet of HTML to indent a group, to show hierarchy.
+
+=cut
+
 sub walkGroups {
 	my $session = shift;
         my $parentId = shift;
@@ -206,9 +274,23 @@ sub walkGroups {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_addGroupsToGroupSave 
+
+Process the addGroupsToGroup form.  Returns adminOnly unless the
+current user canEdit this group.  Group GUIDs are passed in via the
+form variable C<groups>.  Returns the user to the manageGroupsInGroup
+screen.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_addGroupsToGroupSave {
 	my $session = shift;
-	return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
+	return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")) && $session->form->validToken);
 	my $group = WebGUI::Group->new($session,$session->form->process("gid"));
 	my @groups = $session->form->group('groups');
 	$group->addGroups(\@groups);
@@ -216,9 +298,23 @@ sub www_addGroupsToGroupSave {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_addUsersToGroupSave 
+
+Process the addUsersToGroup form.  Returns adminOnly unless the
+current user canEdit this group.  User GUIDs are passed in via the
+form variable C<users>.  Returns the user to the manageUsersInGroup
+screen.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_addUsersToGroupSave {
 	my $session = shift;
-        return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
+        return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")) && $session->form->validToken);
         my @users = $session->form->selectList('users');
 	my $group = WebGUI::Group->new($session,$session->form->process("gid"));
 	$group->addUsers(\@users);
@@ -226,22 +322,48 @@ sub www_addUsersToGroupSave {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_autoAddToGroup 
+
+Web facing method for users to automatically add themselves to a group.
+Returns insufficient if the user is Visitor.  If the group exists and
+has autoAdd set, adds the current user to the group.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_autoAddToGroup {
 	my $session = shift;
         return WebGUI::AdminConsole->new($session,"groups")->render($session->privilege->insufficient()) unless ($session->user->userId ne 1);
 	my $group = WebGUI::Group->new($session,$session->form->process("groupId"));
-	if ($group->autoAdd) {
+	if ($group && $group->autoAdd) {
 		$group->addUsers([$session->user->userId],[$session->form->process("groupId")]);
 	}
 	return "";
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_autoDeleteFromGroup 
+
+Web facing method for users to automatically remove themselves from a group.
+Returns insufficient if the user is Visitor.  If the group exists and
+has autoDelete set, deletes the current user from the group.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_autoDeleteFromGroup {
 	my $session = shift;
         return WebGUI::AdminConsole->new($session,"groups")->render($session->privilege->insufficient()) unless ($session->user->userId ne 1);
 	my $group = WebGUI::Group->new($session,$session->form->process("groupId"));
-	if ($group->autoDelete) {
+	if ($group && $group->autoDelete) {
 		$group->deleteUsers([$session->user->userId],[$session->form->process("groupId")]);
 	}
 	return "";
@@ -255,18 +377,36 @@ Delete's the group specified by id, in the form variable gid.  Groups 1-17
 are reserved for WebGUI internal groups and are not allowed to be deleted.
 Returns you to www_listGroups when done.
 
+=head3 $session
+
+A WebGUI::Session object
+
 =cut
 
 sub www_deleteGroup {
     my $session = shift;
-    return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
-    return $session->privilege->vitalComponent() if (isIn($session->form->process("gid"), qw(1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17)));
-    my $g = WebGUI::Group->new($session,$session->form->process("gid"));
+    my $gid = $session->form->process("gid");
+    return $session->privilege->adminOnly() unless (canEditGroup($session, $gid));
+    return $session->privilege->vitalComponent() if WebGUI::Group->vitalGroup($gid);
+    my $g = WebGUI::Group->new($session,$gid);
     $g->delete;
     return www_listGroups($session);
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_deleteGroupGrouping 
+
+Web facing method for deleting a group from a group.  Returns adminOnly unless
+the current user canEdit this group.  The group GUID to be deleted is passed in
+via the form variable C<delete>.  Returns the user to the manageGroupsInGroup screen.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_deleteGroupGrouping {
 	my $session = shift;
 	return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
@@ -279,16 +419,19 @@ sub www_deleteGroupGrouping {
 
 =head2 www_deleteGrouping ( )
 
-Deletes a set of users from a set of groups.  
-The user and group lists are expected to
-be found in form fields names uid and gid, respectively.  Visitors are not allowed to
-perform this operation.
+Deletes a set of users from a set of groups.  The user and group lists are expected
+to be found in form fields names uid and gid, respectively.  Visitors are not
+allowed to perform this operation.
+
+=head3 $session
+
+A WebGUI::Session object
 
 =cut
 
 sub www_deleteGrouping {
 	my $session = shift;
-        return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
+        return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")) && $session->form->validToken);
         if (($session->user->userId eq $session->form->process("uid") || $session->form->process("uid") eq '3') && $session->form->process("gid") eq '3') {
                 return $session->privilege->vitalComponent();
         }
@@ -303,6 +446,19 @@ sub www_deleteGrouping {
                                                                                                                                                        
 
 #-------------------------------------------------------------------
+
+=head2 www_editGroup 
+
+Renders a form to add or edit groups.  Returns adminOnly unless the
+current user canEdit this group.  The group GUID is passed in via
+the form variable C<gid>.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_editGroup {
 	my $session = shift;
 	return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
@@ -481,11 +637,26 @@ sub www_editGroup {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_editGroupSave 
+
+Process the editGroup form.  Returns adminOnly unless the current
+user canEdit this group.  The GUID of the group is passed in via the
+form variable C<gid>.
+
+Returns the user to the list of groups.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_editGroupSave {
     my $session = shift;
     my $gid = $session->form->process("gid");
     return $session->privilege->adminOnly
-        unless canEditGroup($session, $gid);
+        unless canEditGroup($session, $gid) && $session->form->validToken;
     my $g = WebGUI::Group->new($session, $gid);
     # We don't want them to use an existing name.  If needed, we'll add a number to the name to keep it unique.
     my $groupName = $session->form->process("groupName");
@@ -523,6 +694,20 @@ sub www_editGroupSave {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_editGrouping 
+
+Renders a form for the properties of a user being in a group, namely
+their expire date and whether or not they are an admin for this
+group.  The userId and groupId are passed in via the C<uid> and C<gid>
+form variables, respectively.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_editGrouping {
     my $session = shift;
     my $uid = $session->form->process('uid');
@@ -573,9 +758,22 @@ sub www_editGrouping {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_editGroupingSave 
+
+Process the editGrouping form.  Returns adminOnly unless the current
+user can edit this group.  Returns the user to the manageUsersInGroup
+screen.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_editGroupingSave {
 	my $session = shift;
-	return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
+	return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")) && $session->form->validToken);
 	my $group = WebGUI::Group->new($session,$session->form->process("gid"));
         $group->userGroupExpireDate($session->form->process("uid"),$session->datetime->setToEpoch($session->form->process("expireDate")));
         $group->userIsAdmin($session->form->process("uid"),$session->form->process("groupAdmin"));
@@ -583,6 +781,18 @@ sub www_editGroupingSave {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_emailGroup 
+
+Render a form where an email can be sent to all members of a group.
+Returns adminOnly unless the current user canEdit this group.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_emailGroup {
 	my $session = shift;
 	return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
@@ -628,9 +838,21 @@ sub www_emailGroup {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_emailGroupSend 
+
+Processes the emailGroup form.  Returns adminOnly unless the current
+user canEdit this group.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_emailGroupSend {
 	my $session = shift;
-	return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
+	return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")) && $session->form->validToken);
 	my $mail = WebGUI::Mail::Send->create($session, {toGroup=>$session->form->process("gid"),subject=>$session->form->process("subject"),from=>$session->form->process("from")});
 	$mail->addHtml($session->form->process("message","HTMLArea"));
 	$mail->addFooter;
@@ -640,6 +862,19 @@ sub www_emailGroupSend {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_listGroups 
+
+Render a paginated list of all groups.  If more than 250 groups exist,
+then it will show only a search form.  The list of groups will only
+show that groups that a user canEdit.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_listGroups {
 	my $session = shift;
 	my $i18n = WebGUI::International->new($session);
@@ -703,6 +938,18 @@ sub www_listGroups {
 
 
 #-------------------------------------------------------------------
+
+=head2 www_manageGroupsInGroup 
+
+Render a list of all groups in this group, by hierarchy and a form
+for adding and removing groups.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_manageGroupsInGroup {
 	my $session = shift;
     my $i18n = WebGUI::International->new($session);
@@ -748,6 +995,19 @@ sub www_manageGroupsInGroup {
 }
 
 #-------------------------------------------------------------------
+
+=head2 www_manageUsersInGroup
+
+Renders a form that displays all users in this group, as well as a
+list of users that can be added to this group.  Returns adminOnly
+unless the current user canEdit this group.
+
+=head3 $session
+
+A WebGUI::Session object
+
+=cut
+
 sub www_manageUsersInGroup {
 	my $session = shift;
     return $session->privilege->adminOnly() unless (canEditGroup($session,$session->form->process("gid")));
@@ -775,7 +1035,6 @@ sub www_manageUsersInGroup {
 				name=>"uid",
 				value=>$row->{userId}
 				})
-                        .$session->icon->delete('op=deleteGrouping;uid='.$row->{userId}.';gid='.$session->form->process("gid"))
                         .$session->icon->edit('op=editGrouping;uid='.$row->{userId}.';gid='.$session->form->process("gid"))
                         .'</td>';
                 $output .= '<td class="tableData"><a href="'.$session->url->page('op=editUser;uid='.$row->{userId}).'">'.$row->{username}.'</a></td>';
@@ -828,7 +1087,5 @@ sub www_manageUsersInGroup {
 	$output .= $f->print;
         return _submenu($session,$output,'88');
 }
-
-
 
 1;
